Data Analysis Agent (LangGraph)/requirements.txt
pandas
numpy
scipy
fuzzywuzzy
python-dotenv
pycountry
langgraph